backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
anthropic>=0.34.0
pydantic>=2.10.0
python-dotenv>=1.0.1
